+++ src/vs/sessions/contrib/changes/browser/changesView.contribution.ts
@@ -11,8 +11,6 @@ import { registerIcon } from '../../../../platform/theme/common/iconRegistry.js'
 import { registerWorkbenchContribution2, WorkbenchPhase } from '../../../../workbench/common/contributions.js';
 import { IViewContainersRegistry, ViewContainerLocation, IViewsRegistry, Extensions as ViewContainerExtensions, WindowVisibility } from '../../../../workbench/common/views.js';
 import { CHANGES_VIEW_CONTAINER_ID, CHANGES_VIEW_ID, ChangesViewPane, ChangesViewPaneContainer } from './changesView.js';
-import './changesViewActions.js';
-import './fixCIChecksAction.js';
 import { ChangesTitleBarContribution } from './changesTitleBarWidget.js';
 
 const changesViewIcon = registerIcon('changes-view-icon', Codicon.gitCompare, localize2('changesViewIcon', 'View icon for the Changes view.').value);
