+++ src/vs/sessions/contrib/changes/browser/checksWidget.ts
@@ -23,7 +23,7 @@ import { DEFAULT_LABELS_CONTAINER, IResourceLabel, ResourceLabels } from '../../
 import { ActionBar } from '../../../../base/browser/ui/actionbar/actionbar.js';
 import { GitHubCheckConclusion, GitHubCheckStatus, IGitHubCICheck } from '../../github/common/types.js';
 import { GitHubPullRequestCIModel, parseWorkflowRunId } from '../../github/browser/models/githubPullRequestCIModel.js';
-import { CICheckGroup, buildFixChecksPrompt, getCheckGroup, getCheckStateLabel, getFailedChecks } from './fixCIChecksAction.js';
+import { CICheckGroup, buildFixChecksPrompt, getCheckGroup, getCheckStateLabel, getFailedChecks } from './checksActions.js';
 
 const $ = dom.$;
 
@@ -230,7 +230,7 @@ export class CIStatusWidget extends Disposable {
 		this._chevronNode.classList.add(...ThemeIcon.asClassNameArray(Codicon.chevronDown));
 
 		this._headerNode.setAttribute('role', 'button');
-		this._headerNode.setAttribute('aria-label', localize('ci.toggleChecks', "Toggle PR Checks"));
+		this._headerNode.setAttribute('aria-label', localize('ci.toggleChecks', "Toggle Checks"));
 		this._headerNode.setAttribute('aria-expanded', 'true');
 		this._headerNode.tabIndex = 0;
 
@@ -276,16 +276,11 @@ export class CIStatusWidget extends Disposable {
 		this._bodyNode.appendChild(this._list.getHTMLElement());
 	}
 
-	/**
-	 * Bind to a CI model. When `ciModel` is undefined, the widget hides.
-	 * Returns a disposable that stops observation.
-	 */
-	bind(ciModel: IObservable<GitHubPullRequestCIModel | undefined>, sessionResource: IObservable<URI | undefined>): IDisposable {
+	setInput(input: IObservable<GitHubPullRequestCIModel | undefined>): IDisposable {
 		return autorun(reader => {
-			const model = ciModel.read(reader);
-			this._sessionResource = sessionResource.read(reader);
-			this._model = model;
-			if (!model) {
+			this._model = input.read(reader);
+
+			if (!this._model) {
 				this._checkCount = 0;
 				this._setCollapsed(false);
 				this._renderBody([]);
@@ -295,7 +290,7 @@ export class CIStatusWidget extends Disposable {
 				return;
 			}
 
-			const checks = model.checks.read(reader);
+			const checks = this._model.checks.read(reader);
 
 			if (checks.length === 0) {
 				this._checkCount = 0;
